Day0925 End || Attack() 함수 정리하는 거 부터 계속

+++ Project_1_Kim_WooBin/Enemy.cpp
@@ -1,7 +1,7 @@
 #include "Enemy.h"
 #include <random>
 
-void AEnemy::Attack(Actor* InTarget, bool InIsBlank ,int InProbability) // GameManager에서 GetLiveBulletProbability();로 받아야함
+void AEnemy::Attack(Actor* InTarget, bool InIsBlank ,float InProbability) // GameManager에서 GetLiveBulletProbability();로 받아야함
 {
 	if (InIsBlank == false)
 	{
@@ -38,18 +38,3 @@ void AEnemy::Attack(Actor* InTarget, bool InIsBlank ,int InProbability) // GameM
 
 	
 }
-
-//void AEnemy::IsShoot(Actor* InTarget)
-//{
-//	// 1/3확률로 본인에게 발포, 2/3확률로 플레이어에게 발포
-//	int Possibility = rand() % 3;
-//	printf("Possibility : %d", Possibility);
-//	if (Possibility == 0 || Possibility == 1)
-//	{
-//		Attack(InTarget);
-//	}
-//	else
-//	{
-//		Attack(this);
-//	}
-//}
